+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 ### Changed
 ### Fixed
+- RED.util.evaluateJSONataExpression with callback. (LAT-160)
 ### Dependencies
 #### Updated